+++ .github/workflows/scripts/docs/templates/meta.js
@@ -1,0 +1,4 @@
+window.META = {
+  createdOn: "{{ now }}",
+  version: "{{ version }}"
+};