+++ docs/meta.js
@@ -1,0 +1,4 @@
+window.META = {
+  createdOn: "2026-03-13 22:03:59",
+  version: "0.3.14-471-g744b418"
+};